go: preserve nil structs on decode
When decoding a struct with struct fields, the generated `decode()`
method would replace the `nil` fields with an empty new struct. Since
`nil` fields are stored in the original message with zero bytes
(`0x00`), we can check for it before attempting to decode the field into
an intialized zero value.

Signed-off-by: Luiz Aoqui <luizaoqui@loopholelabs.io>